// server/src/enums/enums.ts
export enum Figure {
  X = 'X',
  O = 'O',
  NONE = 'NONE',
}

export enum Player {
  PLAYER_1 = 'PLAYER_1',
  PLAYER_2 = 'PLAYER_2',
}

export enum GameStatus {
  PLAYER_1_WON = 'PLAYER_1_WON',
  PLAYER_2_WON = 'PLAYER_2_WON',
  DRAW = 'DRAW',
  PENDING = 'PENDING',
  IN_PROGRESS = 'IN_PROGRESS',
}